Removed some new warnings.

+++ emu/random_xsb.c
@@ -48,17 +48,17 @@ static short IX = 6293;
 static short IY = 21877;
 static short IZ = 7943;
 
-static Float TX = 1.0/30269.0;
-static Float TY = 1.0/30307.0;
-static Float TZ = 1.0/30323.0;
+static double TX = 1.0/30269.0;
+static double TY = 1.0/30307.0;
+static double TZ = 1.0/30323.0;
 
 /*
  * Returns a float number within the range [0.0, 1.0) in reg 2.
  * ret_random() returns FALSE if there is an error, TRUE if everything is OK.
  */
 int ret_random() {
-  long X, Y, Z;
-  Float T;
+  short X, Y, Z;
+  double T;
   Cell term;
 
   X = (IX*171) % 30269;
